requirements.txt
django==2.0
djangorestframework==3.8.2
requests
psycopg2
cerberus